Assignment of unsigned received bytes in expression is always false.
When compiling with clang/gcc the following warning is emitted:

    warning: comparison of constant -1 with expression of type 'uint16_t'
    (aka 'unsigned short') is always false
    [-Wtautological-constant-out-of-range-compare]

Changing received_bytes to match the return type of recvfrom() resolves
this warning and also requires changing the format statements within
the printf() calls.

+++ srf-ip-conn/client-demo/client-sock.h
@@ -32,7 +32,7 @@ DEALINGS IN THE SOFTWARE.
 
 typedef struct {
 	uint8_t buf[sizeof(srf_ip_conn_packet_t)];
-	uint16_t received_bytes;
+	ssize_t received_bytes;
 } client_sock_received_packet_t;
 
 extern client_sock_received_packet_t client_sock_received_packet;
